Fix cargo clippy warnings
- Remove unused imports (AsyncBlockSource, SyncBlockSource, BlockSourceByteStream, ProductBlockSource)
- Add #[allow(dead_code)] for async trait methods that are part of the API but not used internally
- Remove needless borrow in block_writer.rs
- Remove unnecessary cast in block_source_byte_stream.rs

+++ src/byte_stream/block_source_byte_stream.rs
@@ -139,7 +139,7 @@ impl Drop for Buffer {
     fn drop(&mut self) {
         // Check the buffer_size invariant. If it's correct now, it probably was correct all the time.
         let actual_buffer_size: usize =
-            self.buffer.iter().map(|b| b.len()).sum::<usize>() - self.current_pos as usize;
+            self.buffer.iter().map(|b| b.len()).sum::<usize>() - self.current_pos;
         assert_eq!(
             self.buffer_size, actual_buffer_size,
             "Miscalculated self.buffer_size"
